Move ServiceBusBuildTimeConfig to deployment module
